requirements.txt
crawl4ai>=0.5.0
supabase>=2.4.0
python-dotenv>=1.0.1
pydantic>=2.6.0
httpx>=0.27.0
